// src/Model/ParentReport.model.tsx
export interface ParentReport {
    parentReportId: string;
    studentId: string;
    strengthArea: string;
    birthProcessAndEarlyDevelopment: string;
    weaning: string;
    preSchoolEducation: string;
    ParamedicalSupport: string;
    PreSchoolDifficulties: string;
    elementaryEducation: string;
    readingAndWritingDevelopment: string;
    understandingInstructions: string;
    elementarySchoolDifficulties: string;
    academicDifficulties: string;
    socialAndEmotionalDifficulties: string;
    wasWithoutFramework: boolean;
    currentAcademicStatus: string;
    currentReadingGap: string;
    currentWritingGap: string;
    currentMotivationGap: string;
    currentSocialGap: string;
    currentEmotionalGap: string;
    currentCulturalAndLeisureGap: string;

}